Refactor previews.
