+++ app/build.gradle
@@ -34,7 +34,7 @@ plugins {
 }
 
 android {
-    compileSdkVersion 35
+    compileSdk 36
     defaultConfig {
         applicationId "com.exercise.davismiyashiro.popularmovies"
         minSdkVersion 21
@@ -78,11 +78,8 @@ android {
 
     useLibrary 'android.test.runner'
     namespace 'com.exercise.davismiyashiro.popularmovies'
-}
-
-tasks.withType(KotlinCompile).configureEach {
     kotlinOptions {
-        jvmTarget = "17"
+        jvmTarget = '17'
     }
 }
 
@@ -112,8 +109,8 @@ dependencies {
     implementation 'com.squareup.moshi:moshi:1.15.0'
     ksp 'com.squareup.moshi:moshi-kotlin-codegen:1.15.0'
 
-    // implementation 'com.squareup.picasso:picasso:2.71828' // Picasso removed
-    implementation("io.coil-kt:coil-compose:2.6.0") // Coil added
+    implementation("io.coil-kt.coil3:coil-compose:3.3.0")
+    implementation("io.coil-kt.coil3:coil-network-okhttp:3.3.0")
 
     implementation 'com.jakewharton.timber:timber:4.7.1'
 
@@ -129,14 +126,15 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.9.2"
 
     // Jetpack Compose
-    def composeBom = platform('androidx.compose:compose-bom:2024.06.00')
+    def composeBom = platform('androidx.compose:compose-bom:2025.08.00')
     implementation composeBom
     androidTestImplementation composeBom
     implementation 'androidx.compose.ui:ui'
     implementation 'androidx.compose.material3:material3'
     debugImplementation 'androidx.compose.ui:ui-tooling'
     implementation 'androidx.compose.ui:ui-tooling-preview'
-    implementation 'androidx.activity:activity-compose:1.9.0'
+    implementation 'androidx.compose.ui:ui-graphics'
+    implementation 'androidx.activity:activity-compose:1.10.1'
 
     testImplementation 'org.mockito:mockito-core:5.18.0'
     testImplementation 'com.squareup.okhttp3:mockwebserver:5.1.0'
