+++ app/src/main/java/com/exercise/davismiyashiro/popularmovies/moviedetails/MovieDetailsActivity.kt
@@ -65,12 +65,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import androidx.lifecycle.ViewModelProvider
-import coil.compose.SubcomposeAsyncImage
+import coil3.compose.SubcomposeAsyncImage
 import com.exercise.davismiyashiro.popularmovies.App
 import com.exercise.davismiyashiro.popularmovies.R
 import com.exercise.davismiyashiro.popularmovies.data.Review
@@ -201,8 +201,13 @@ fun MovieDetailsContent(
         // Movie Backdrop and Favorite Icon
         item {
             Box(modifier = Modifier.fillMaxWidth()) {
-                ImagePlaceholder( // Updated to use Coil's AsyncImage
-                    url = IMG_BASE_URL + movieDetails.backdropPath,
+                val headerImage = if (LocalInspectionMode.current) {
+                    R.drawable.header
+                } else {
+                    movieDetails.backdropPath
+                }
+                ImagePlaceholder(
+                    model = headerImage,
                     contentDescription = "Movie Backdrop",
                     modifier = Modifier
                         .fillMaxWidth()
@@ -231,12 +236,17 @@ fun MovieDetailsContent(
             Row(
                 verticalAlignment = Alignment.Top
             ) {
-                ImagePlaceholder( // Updated to use Coil's AsyncImage
-                    url = IMG_BASE_URL + movieDetails.posterPath,
+                val poster = if (LocalInspectionMode.current) {
+                    R.drawable.poster
+                } else {
+                    movieDetails.posterPath
+                }
+                ImagePlaceholder(
+                    model = poster,
                     contentDescription = "Movie Poster",
                     modifier = Modifier
                         .width(120.dp)
-                        .height(180.dp) // Maintain aspect ratio roughly
+                        .height(180.dp)
                 )
                 Spacer(modifier = Modifier.width(16.dp))
                 Column {
@@ -309,12 +319,18 @@ fun MovieDetailsContent(
 }
 
 @Composable
-fun ImagePlaceholder(url: String, contentDescription: String?, modifier: Modifier = Modifier) {
+fun ImagePlaceholder(model: Any, contentDescription: String?, modifier: Modifier = Modifier) {
     SubcomposeAsyncImage(
         modifier = modifier,
-        model = url,
+        model = model,
         contentDescription = contentDescription,
-        contentScale = ContentScale.Crop
+        contentScale = ContentScale.Crop,
+        loading = {
+            CircularProgressIndicator()
+        },
+        error = {
+
+        }
     )
 }
 
@@ -360,135 +376,4 @@ fun ReviewItem(review: Review, onClick: () -> Unit) {
 @Composable
 fun stringResource(@StringRes id: Int, vararg formatArgs: Any): String {
     return LocalContext.current.getString(id, *formatArgs)
-}
-
-// --- Previews ---
-
-@Preview(showBackground = true)
-@Composable
-fun ImagePlaceholderPreview() {
-    MaterialTheme {
-        ImagePlaceholder(
-            url = "https://example.com/image.jpg",
-            contentDescription = "Sample Image"
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun TrailerItemPreview() {
-    MaterialTheme {
-        val sampleTrailer = Trailer(
-            id = "1",
-            key = "abcdef",
-            name = "Official Trailer",
-            site = "YouTube",
-            size = 1080,
-            type = "Trailer",
-            iso6391 = "iso6391",
-            iso31661 = "iso6391"
-        )
-        TrailerItem(trailer = sampleTrailer, onClick = {})
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ReviewItemPreview() {
-    MaterialTheme {
-        val sampleReview = Review(
-            id = "1",
-            author = "John Doe",
-            content = "This is a great movie! ".repeat(10),
-            url = ""
-        )
-        ReviewItem(review = sampleReview, onClick = {})
-    }
-}
-
-@Preview(showBackground = true, name = "MovieDetailsContent - Populated")
-@Composable
-fun MovieDetailsContentPopulatedPreview() {
-    val sampleMovieDetails = MovieDetailsObservable(
-        id = 1,
-        title = "Awesome Movie Title",
-        posterPath = "/poster.jpg",
-        overview = "This is a really awesome movie that you should definitely watch. ".repeat(5),
-        releaseDate = "2024-01-01",
-        voteAverage = 8.5,
-        backdropPath = "/backdrop.jpg"
-    )
-    val sampleTrailers = listOf(
-        Trailer(
-            id = "1",
-            key = "key1",
-            name = "Trailer 1",
-            site = "YouTube",
-            type = "Trailer",
-            size = 1080,
-            iso6391 = "iso6391",
-            iso31661 = "iso6391"
-        ),
-        Trailer(
-            id = "2",
-            key = "key2",
-            name = "Trailer 2 - Extended Cut",
-            site = "YouTube",
-            type = "Trailer",
-            size = 1080,
-            iso6391 = "iso6391",
-            iso31661 = "iso6391"
-        )
-    )
-    val sampleReviews = listOf(
-        Review(
-            id = "1",
-            author = "Jane Critic",
-            content = "A cinematic masterpiece! ".repeat(3),
-            url = ""
-        ),
-        Review(
-            id = "2",
-            author = "Bob Reviewer",
-            content = "Simply stunning visuals and compelling story. ".repeat(3),
-            url = ""
-        )
-    )
-    MaterialTheme {
-        MovieDetailsContent(
-            movieDetails = sampleMovieDetails,
-            trailers = sampleTrailers,
-            reviews = sampleReviews,
-            isFavorite = true,
-            onFavoriteToggle = {},
-            onTrailerClick = {},
-            onReviewClick = {}
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "MovieDetailsContent - Empty")
-@Composable
-fun MovieDetailsContentEmptyPreview() {
-    val sampleMovieDetails = MovieDetailsObservable(
-        id = 1,
-        title = "Awesome Movie Title",
-        posterPath = "/poster.jpg",
-        overview = "This is a really awesome movie that you should definitely watch. ".repeat(5),
-        releaseDate = "2024-01-01",
-        voteAverage = 8.5,
-        backdropPath = "/backdrop.jpg"
-    )
-    MaterialTheme {
-        MovieDetailsContent(
-            movieDetails = sampleMovieDetails,
-            trailers = emptyList(),
-            reviews = emptyList(),
-            isFavorite = false,
-            onFavoriteToggle = {},
-            onTrailerClick = {},
-            onReviewClick = {}
-        )
-    }
 }
