+++ app/src/main/java/com/exercise/davismiyashiro/popularmovies/movies/MovieListAdapter.kt
@@ -27,10 +27,8 @@ package com.exercise.davismiyashiro.popularmovies.movies
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import coil.load
-import com.exercise.davismiyashiro.popularmovies.R
+import coil3.load
 import com.exercise.davismiyashiro.popularmovies.databinding.MovieListItemBinding
-import com.exercise.davismiyashiro.popularmovies.moviedetails.IMG_BASE_URL
 import com.exercise.davismiyashiro.popularmovies.moviedetails.MovieDetailsObservable
 
 /**
@@ -72,10 +70,7 @@ class MovieListAdapter(
 
         fun bind(movieDetails: MovieDetailsObservable) {
             if (movieDetails.posterPath.isNotEmpty()) {
-                binding.imgUrl.load(IMG_BASE_URL + movieDetails.posterPath) {
-                    placeholder(android.R.drawable.progress_indeterminate_horizontal)
-                    error(android.R.drawable.stat_notify_error)
-                }
+                binding.imgUrl.load(movieDetails.posterPath)
             }
             binding.imgUrl.setOnClickListener { mClickListener.getMovieClicked(movieDetails) }
         }
